fixed all segmentation faults, all maintests and unit tests pass

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: ksiziva <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/24 15:09:58 by ksiziva           #+#    #+#             */
-/*   Updated: 2018/05/31 10:28:03 by ksiziva          ###   ########.fr       */
+/*   Updated: 2018/06/16 20:14:00 by kaysiz           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	ft_memset(s, 0, n);
 }
